requirements.txt
python-dotenv
langchain
ipykernel
langchain-community
langchain-core
langchain-groq
langchain-google-genai
langchain_huggingface
sentence-transformers
openai
faiss-cpu
pypdf
streamlit
bs4
pymupdf
langchain-text-splitters
duckdb
pandas
langchain_chroma